% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetCelltypes}
\alias{GetCelltypes}
\title{GetCelltypes}
\usage{
GetCelltypes(organism, organ)
}
\arguments{
\item{organism}{The organism you would like to query}

\item{organ}{The organ you would like to query}
}
\value{
An array of available cell types from that organism and organ
}
\description{
GetCelltypes
}
\examples{
GetCelltypes("h_sapiens", "Lung")
}
