% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetDataSources}
\alias{GetDataSources}
\title{GetDataSources}
\usage{
GetDataSources()
}
\value{
A list containing information about the cell atlases used as data sources
}
\description{
Get information about the cell atlases used as data sources for the approximations.
}
\examples{
GetDataSources()
}
