% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetFeatures}
\alias{GetFeatures}
\title{GetFeatures}
\usage{
GetFeatures(organism)
}
\arguments{
\item{organism}{The organism you would like to query}
}
\value{
An array of available features (genes) from that organism
}
\description{
GetFeatures
}
\examples{
GetFeatures("h_sapiens")
}
