% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetFractionDetected}
\alias{GetFractionDetected}
\title{GetFractionDetected}
\usage{
GetFractionDetected(organism, organ, features)
}
\arguments{
\item{organism}{The organism you would like to query}

\item{organ}{The organ you would like to query}

\item{features}{The features (genes) you would like to query}
}
\value{
A data.frame of fraction of expressing cells by cell type in that organism and organ
}
\description{
Get the fraction of cells expressing specified features across cell types in a given organism and organ.
}
\examples{
GetFractionDetected("h_sapiens", "Lung", c("COL1A1", "PTPRC"))
}
