% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetHighestMeasurement}
\alias{GetHighestMeasurement}
\title{GetHighestMeasurement}
\usage{
GetHighestMeasurement(organism, feature, number)
}
\arguments{
\item{organism}{The organism you would like to query}

\item{feature}{The feature you would like to query}

\item{number}{The number of highest expressors you would like to get}
}
\value{
A dataframe of cell types, organs, and averages for the
cell types with the highest measurement for that feature
}
\description{
Get the cell types with the highest expression of a specified feature in a given organism.
}
\examples{
GetHighestMeasurement("h_sapiens", "PTPRC", 5)
}
