% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetMarkers}
\alias{GetMarkers}
\title{GetMarkers}
\usage{
GetMarkers(organism, organ, cell_type, number)
}
\arguments{
\item{organism}{The organism you would like to query}

\item{organ}{The organism you would like to query}

\item{cell_type}{The cell type you would like to query}

\item{number}{The number of markers you would like to get}
}
\value{
An array of markers for that cell type in that organism and organ
}
\description{
Get marker genes for a specified cell type in a given organism and organ.
}
\examples{
GetMarkers("h_sapiens", "Lung", "fibroblast", 5)
}
