% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetOrganisms}
\alias{GetOrganisms}
\title{GetOrganisms}
\usage{
GetOrganisms()
}
\value{
An array of available organisms
}
\description{
Get a list of organisms available for querying in the atlasapprox api.
}
\examples{
GetOrganisms()
}
