% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetOrgans}
\alias{GetOrgans}
\title{GetOrgans}
\usage{
GetOrgans(organism)
}
\arguments{
\item{organism}{The organism you would like to query}
}
\value{
An array of available organs from that organism
}
\description{
GetOrgans
}
\examples{
GetOrgans("h_sapiens") 
}
