% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetSimilarFeatures}
\alias{GetSimilarFeatures}
\title{GetSimilarFeatures}
\usage{
GetSimilarFeatures(organism, organ, feature, number, method)
}
\arguments{
\item{organism}{The organism you would like to query}

\item{organ}{The organ you would like to query}

\item{feature}{The feature to find similarities for}

\item{number}{The number of similar features to return}

\item{method}{The method used for the distance computation.
Available methods are: "correlation" (default), "cosine",
"euclidean", "manhattan", "log-euclidean".}
}
\value{
An dataframe of similar features and their distances from the focal feature
according to the method chosen.
}
\description{
GetSimilarFeatures
}
\examples{
GetSimilarFeatures("h_sapiens", "lung", "PTPRC", 5, "correlation")
}
