% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetAverage}
\alias{GetAverage}
\title{GetAverage}
\usage{
GetAverage(organism, organ, features)
}
\arguments{
\item{organism}{The organism you would like to query}

\item{organ}{The organism you would like to query}

\item{features}{The features (genes) you would like to query}
}
\value{
A data.frame of average gene expression by cell type in that organism and organ
}
\description{
Get the average gene expression for specified features across cell types in a given organism and organ.
}
\examples{
GetAverage("h_sapiens", "Lung", c("COL1A1", "PTPRC"))
}
