% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetDataSources}
\alias{GetDataSources}
\title{GetDataSources}
\usage{
GetDataSources()
}
\value{
The cell atlases used as data sources for the approximations
}
\description{
GetDataSources
}
\examples{
GetDataSources() 
}
