% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasapprox.R
\name{GetHighestMeasurement}
\alias{GetHighestMeasurement}
\title{GetHighestMeasurement}
\usage{
GetHighestMeasurement(organism, feature, number)
}
\arguments{
\item{organism}{The organism you would like to query}

\item{feature}{The feature to check (e.g. gene)}

\item{number}{The number of cell types to return}
}
\value{
An dataframe of cell types, organs, and averages for the
cell types with the highest measurement for that feature
}
\description{
GetHighestMeasurement
}
\examples{
GetHighestMeasurement("h_sapiens", "PTPRC", 5)
}
